Calculated the answers to the multiplications and stored them in operationAnswers.
